refactor(coordinator): extract provider persistence from registry.go
Move the registry<->store persistence bridge (SetStore, LoadStoredProviders,
RestoreProviderState, providerRecordStats, PersistProvider[Throttled],
persistReputation[Throttled], persistProviderNow) into registry/persistence.go.
Same-package move, no behavior change; registry.go 4693 -> 4365 LOC.
registry go test ./registry/... stays green.
